fixes...
- add behat test for lookup;
- add sql data for test;
- add dropdown cascade using new data;
